GymApp/requirements.txt
django
django-extensions
psycopg2-binary
djangorestframework
django-cors-headers